[MLIR][LLVM] Don't use void return type in MLIR APIs.
In the LLVM IR dialect, `LLVMVoidType` is used to model the return type
of LLVM IR functions with no return value. This is inconsistent with
MLIR APIs, which expect a function with no return value to have an
empty return type. Handle this special case in `LLVMFuncOp` to avoid
mismatches between the number of return values and return types between
caller and callee.

The parser of `LLVMFuncOp` is adapted to be symmetric with the printer,
and not allow `!llvm.void` to be expressed explicitly as a return type
in the IR, thus avoiding to expose this implementation detail.

Differential Revision: https://reviews.llvm.org/D141676